Adjusting one int type to long int + corrected doc

+++ src/qiskit_qec/analysis/combinations.h
@@ -23,7 +23,7 @@ class Combinations
   public:
     Combinations(int n_, int k_);
     bool next_combination(std::vector<int> &state);
-    int size(void);
+    long int size(void);
 };
 
 #endif
